pyproject.toml
[project]
name = "galileo-agents"
version = "0.1.0"
description = "AI agents"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "galileo[otel]",
    "python-dotenv",
    "traceloop-sdk",
]

[project.optional-dependencies]
langgraph = [
    "langgraph",
    "langgraph-prebuilt",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "faiss-cpu",
]
crewai = [
    "crewai",
    "crewai-tools",
    "langchain-openai",
    "langchain-community",
    "faiss-cpu",
]
all = [
    "galileo-agents[langgraph]",
    "galileo-agents[crewai]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]

[tool.ruff]
line-length = 120
